Add files via upload
This dashboard indicates the NYPD Arrest Date (Year to Date ). The project focuses on the number of arrests that occurred by a person's, a group's or a place's identification with a particular race, religion, ethnicity, gender, age, disability, ancestry, national origin or sexual orientation.  . 
Arrests made by the responding or investigating officers are recorded as part of the booking process and contain race and ethnicity information for the arrested suspects. This dataset was gotten from Onyx challenge